Refactor: 컴포넌트명 일치
