+++ src/hooks/useHeaderLogic.ts
