+++ src/views/Recipes/RecommendedRecipe/RecommendedRecipe.tsx
@@ -1,0 +1,35 @@
+// import { useEffect, useState } from 'react';
+import RecipeList from '../Recipe/RecipeList.js';
+import RecipePageHeader from '../Recipe/RecipePageHeader.js';
+// import axios from 'axios';
+import fakeData from '../../../fakeData/recipeFake.js';
+import styled from 'styled-components';
+
+interface RecipeLimitProps {
+    limit?: number;
+    page?: string;
+}
+export default function Recommend({ limit, page }: RecipeLimitProps): JSX.Element {
+    // const [recipes, setRecipes] = useState<string[]>([]);
+    // useEffect(() => {
+    //     const fetchRecipes = async () => {
+    //         try {
+    // const response = await axios.get(`${import.meta.env.VITE_BASE_URL}/recipes`);
+    //         } catch (err) {}
+    //     };
+    // }, [recipes]);
+
+    return (
+        <>
+            {page ? <RecipePageTitle>Recommend Recipes</RecipePageTitle> : <RecipePageHeader title="Recommend" />}
+            <RecipeList recipes={fakeData} limit={limit} page={page} />
+        </>
+    );
+}
+
+const RecipePageTitle = styled.h2`
+    margin: 36px 16px 0px;
+    color: #622b18;
+    font-size: 40px;
+    text-align: center;
+`;
